Move ActivitiesDateFilter enum to constants
