+++ packages/components/dashboard/content.tsx
@@ -15,7 +15,7 @@ import { GroupsList } from '@acter/components/dashboard/groups-list'
 import { ActivitiesDateFilters } from '@acter/components/search/molecules/date-filters'
 import { flattenFollowing } from '@acter/lib/acter/flatten-following'
 import { useActivities } from '@acter/lib/activity/use-activities'
-import { ActivitiesDateFilter } from '@acter/lib/api/resolvers/date-filter'
+import { ActivitiesDateFilter } from '@acter/lib/constants'
 import { ActerTypes } from '@acter/lib/constants'
 import { useTranslation } from '@acter/lib/i18n/use-translation'
 import { useUser } from '@acter/lib/user/use-user'
