+++ packages/lib/constants/activity-date-filter.ts
@@ -1,0 +1,5 @@
+export enum ActivitiesDateFilter {
+  ALL = 'ALL',
+  UPCOMING = 'UPCOMING',
+  PAST = 'PAST',
+}
