+++ services/api/resolvers/search/index.ts
@@ -1,13 +1,11 @@
 import { Resolver, Query, Arg, Ctx, registerEnumType, Int } from 'type-graphql'
 
-import {
-  withDateFilterSearch,
-  ActivitiesDateFilter,
-} from '@acter/lib/api/resolvers/date-filter'
+import { withDateFilterSearch } from '@acter/lib/api/resolvers/date-filter'
 import {
   getOrderBy,
   SearchActivitiesSortBy,
 } from '@acter/lib/api/resolvers/get-order-by'
+import { ActivitiesDateFilter } from '@acter/lib/constants'
 import type { ActerGraphQLContext } from '@acter/lib/types/graphql-api'
 import { Acter } from '@acter/schema'
 import { ActerPrivacySettings } from '@acter/schema'
